Refactor: move fncs to separate files, break up large functions
