+++ lynxkite-assistant/src/lynxkite_assistant/assistant.py
@@ -10,6 +10,7 @@ import deepagents
 from deepagents import backends
 from .workspace_backend import WorkspaceBackend
 from lynxkite_core import workspace
+from .instructions import SYSTEM_PROMPT
 
 try:
     from lynxkite_app import crdt
@@ -17,26 +18,6 @@ except ImportError:
     crdt = None  # type: ignore
 
 router = fastapi.APIRouter()
-
-SYSTEM_PROMPT = """
-You are an assistant for the LynxKite no-code AI workflow builder.
-The user sees the workflow in a visual representation. You have access to it as a file in `workspace.py`, which the user does not see.
-Each function call in `workspace.py` corresponds to a box in the visual representation. Boxes can be connected to each other by their inputs and outputs.
-You may change the layout of the boxes in the visual representation by editing `layout.json`. The user does not see this file, but they will see the updated layout in the visual representation.
-When editing the layout, keep in mind that boxes that are connected should be placed closer to each other, as these boxes are connected by arrows in the visual representation.
-Edit this file to implement the user's requests. `workspace.py` must only contain function calls.
-Keyword arguments must be constants or previous results. Positional arguments are not allowed.
-DO NOT REMOVE or edit any existing code or comments! (unless asked explicitly by the user).
-When adding new comments, make sure to add them above the relevant line of code, so they appear above the box they are associated with.
-New boxes can be added by editing `boxes.py`. Follow the existing conventions in `boxes.py` when defining a new box.
-The new box can be used in `workspace.py` by calling the function from `boxes.py`. The functions are available under a custom module name, specified at the beginning of `boxes.py`.
-You must use existing boxes directly in `workspace.py` without adding them to `boxes.py`.
-You can see any errors that occurred in the boxes in `errors.txt`. Before finishing a task you MUST FIX ALL ERRORS in the new boxes.
-If a custom box returns an 'Unknown operation' error message, check if you are using the correct module name for the new box.
-The module name and usage examples are specified at the beginning of `boxes.py`.
-Attempt to fix any errors in the boxes you add, and if you cannot, explain to the user what went wrong and how to fix it.
-If workspace.py is empty, you can still add new boxes by editing boxes.py or using the pre-defined boxes.
-"""
 
 
 class AssistantMessage(pydantic.BaseModel):
