+++ lynxkite-assistant/src/lynxkite_assistant/python_workspace_conversion.py
@@ -4,9 +4,11 @@ Convert a series of Python function calls into a workspace, or the other way.
 
 import ast
 import graphlib
-from lynxkite_core import ops, workspace
 from functools import reduce
 from itertools import groupby
+from lynxkite_core import ops, workspace
+from . import workspace_comments
+from .instructions import WORKSPACE_PROMPT
 
 
 def _get_func_name(func: ast.expr, error_msg: str) -> str:
@@ -26,25 +28,86 @@ def _get_catalog():
     return catalog
 
 
-def _gather_multiline_comments(code: str) -> list[tuple[int, str]]:
-    comments = []
-    comment_lines = []
-    next_line = 0
-    for i, line in enumerate(code.splitlines()):
-        if not line.strip().startswith("#!"):
-            continue
-        if next_line == i:
-            comment_lines.append(line.strip()[2:].strip())
+def add_edge(arg_name, arg_value_list, saved_values, ws, box_id, error_msg):
+    for arg_value in arg_value_list:
+        if isinstance(arg_value, ast.Subscript):
+            assert isinstance(arg_value.value, ast.Name), error_msg
+            assert isinstance(arg_value.slice, (ast.Constant)), error_msg
+            name = arg_value.value.id
+            sourceHandle = arg_value.slice.value
         else:
-            if comment_lines:
-                comments.append(
-                    (next_line - len(comment_lines), "\n".join(comment_lines))
-                )
-            comment_lines = [line.strip()[2:].strip()]
-        next_line = i + 1
-    if comment_lines:
-        comments.append((next_line - len(comment_lines), "\n".join(comment_lines)))
-    return comments
+            name = arg_value.id
+            sourceHandle = "output"
+        assert name in saved_values, (
+            f"{error_msg}\n\nUnknown variable reference: {name}"
+        )
+        src = saved_values[name]
+        ws.add_edge(src, sourceHandle, box_id, arg_name)
+
+
+def parse_args(
+    box_id,
+    kwargs,
+    saved_values,
+    comment_by_cleaned_text,
+    ws,
+    error_msg,
+    func_name,
+    groups,
+):
+    params = {}
+    for arg_name, arg_value in kwargs.items():
+        assert isinstance(
+            arg_value,
+            (ast.Constant, ast.Name, ast.Dict, ast.List, ast.Tuple, ast.Subscript),
+        ), error_msg
+        if isinstance(arg_value, ast.Constant):
+            params[arg_name] = arg_value.value
+        elif isinstance(arg_value, (ast.Name, ast.Subscript)):
+            add_edge(arg_name, [arg_value], saved_values, ws, box_id, error_msg)
+        elif isinstance(arg_value, ast.Dict):
+            dict_value = {}
+            for key_node, value_node in zip(arg_value.keys, arg_value.values):
+                assert isinstance(key_node, ast.Constant) and isinstance(
+                    key_node.value, str
+                ), error_msg
+                assert isinstance(value_node, ast.Constant), error_msg
+                dict_value[key_node.value] = value_node.value
+            params[arg_name] = dict_value
+        elif isinstance(arg_value, ast.List):
+            if func_name == "lynxkite_core.ops.group" and all(
+                isinstance(item, (ast.Name, ast.Constant)) for item in arg_value.elts
+            ):
+                boxes = {
+                    saved_values[item.id]
+                    for item in arg_value.elts
+                    if isinstance(item, ast.Name)
+                }
+                comments = {
+                    comment_by_cleaned_text[
+                        item.value.replace(" ", "").replace("\n", "")
+                    ]
+                    for item in arg_value.elts
+                    if isinstance(item, ast.Constant) and isinstance(item.value, str)
+                }
+                groups[box_id] = boxes.union(comments)
+            elif all(
+                isinstance(item, (ast.Name, ast.Subscript)) for item in arg_value.elts
+            ):
+                add_edge(arg_name, arg_value.elts, saved_values, ws, box_id, error_msg)
+            else:
+                list_value = []
+                for item in arg_value.elts:
+                    assert isinstance(item, ast.Constant), error_msg
+                    list_value.append(item.value)
+                params[arg_name] = list_value
+        elif isinstance(arg_value, ast.Tuple):
+            tuple_value = []
+            for item in arg_value.elts:
+                assert isinstance(item, ast.Constant), error_msg
+                tuple_value.append(item.value)
+            params[arg_name] = tuple(tuple_value)
+    return params
 
 
 def python_to_workspace(
@@ -55,7 +118,8 @@ def python_to_workspace(
     ws = workspace.Workspace()
     saved_values = {}
     comment_by_cleaned_text = {}
-    for line, text in _gather_multiline_comments(code):
+    # parse comments separately - they do not appear in the AST
+    for line, text in workspace_comments.gather_multiline_comments(code):
         comment_by_cleaned_text[text.replace(" ", "").replace("\n", "")] = (
             f"comment on line {line}"
         )
@@ -101,78 +165,16 @@ def python_to_workspace(
                 f"{error_msg}\n\n**kwargs expansion is not supported."
             )
             kwargs[kw.arg] = kw.value
-        params = {}
-
-        def add_edge(arg_name, arg_value_list):
-            for arg_value in arg_value_list:
-                if isinstance(arg_value, ast.Subscript):
-                    assert isinstance(arg_value.value, ast.Name), error_msg
-                    assert isinstance(arg_value.slice, (ast.Constant)), error_msg
-                    name = arg_value.value.id
-                    sourceHandle = arg_value.slice.value
-                else:
-                    name = arg_value.id
-                    sourceHandle = "output"
-                assert name in saved_values, (
-                    f"{error_msg}\n\nUnknown variable reference: {name}"
-                )
-                src = saved_values[name]
-                ws.add_edge(src, sourceHandle, box_id, arg_name)
-
-        for arg_name, arg_value in kwargs.items():
-            assert isinstance(
-                arg_value,
-                (ast.Constant, ast.Name, ast.Dict, ast.List, ast.Tuple, ast.Subscript),
-            ), error_msg
-            if isinstance(arg_value, ast.Constant):
-                params[arg_name] = arg_value.value
-            elif isinstance(arg_value, (ast.Name, ast.Subscript)):
-                add_edge(arg_name, [arg_value])
-            elif isinstance(arg_value, ast.Dict):
-                dict_value = {}
-                for key_node, value_node in zip(arg_value.keys, arg_value.values):
-                    assert isinstance(key_node, ast.Constant) and isinstance(
-                        key_node.value, str
-                    ), error_msg
-                    assert isinstance(value_node, ast.Constant), error_msg
-                    dict_value[key_node.value] = value_node.value
-                params[arg_name] = dict_value
-            elif isinstance(arg_value, ast.List):
-                if func_name == "lynxkite_core.ops.group" and all(
-                    isinstance(item, (ast.Name, ast.Constant))
-                    for item in arg_value.elts
-                ):
-                    boxes = {
-                        saved_values[item.id]
-                        for item in arg_value.elts
-                        if isinstance(item, ast.Name)
-                    }
-                    comments = {
-                        comment_by_cleaned_text[
-                            item.value.replace(" ", "").replace("\n", "")
-                        ]
-                        for item in arg_value.elts
-                        if isinstance(item, ast.Constant)
-                        and isinstance(item.value, str)
-                    }
-                    groups[box_id] = boxes.union(comments)
-                elif all(
-                    isinstance(item, (ast.Name, ast.Subscript))
-                    for item in arg_value.elts
-                ):
-                    add_edge(arg_name, arg_value.elts)
-                else:
-                    list_value = []
-                    for item in arg_value.elts:
-                        assert isinstance(item, ast.Constant), error_msg
-                        list_value.append(item.value)
-                    params[arg_name] = list_value
-            elif isinstance(arg_value, ast.Tuple):
-                tuple_value = []
-                for item in arg_value.elts:
-                    assert isinstance(item, ast.Constant), error_msg
-                    tuple_value.append(item.value)
-                params[arg_name] = tuple(tuple_value)
+        params = parse_args(
+            box_id,
+            kwargs,
+            saved_values,
+            comment_by_cleaned_text,
+            ws,
+            error_msg,
+            func_name,
+            groups,
+        )
         op = catalog.get(func_name)
         if op:
             box_title = op.name
@@ -206,91 +208,6 @@ def python_to_workspace(
     return ws
 
 
-def describe_schema(schema: dict) -> str:
-    lines = [f"bundle with {len(schema['dataframes'])} dataframe(s)"]
-    for df_name, info in sorted(schema["dataframes"].items()):
-        cols = ", ".join(f"'{c}'" for c in info.get("columns", []))
-        lines.append(f"'{df_name}' with columns {cols}")
-    return "\n# ".join(lines)
-
-
-def compare_dataframe_schemas(prev_schema: dict, new_schema: dict) -> str:
-    changes = []
-    prev_keys = set(prev_schema["dataframes"].keys())
-    new_keys = set(new_schema["dataframes"].keys())
-    all_keys = prev_keys.union(new_keys)
-    if prev_keys.intersection(new_keys) == set():
-        return f"new bundle: {describe_schema(new_schema)}"
-    for key in sorted(all_keys):
-        if key not in prev_schema["dataframes"]:
-            cols = ", ".join(f"'{c}'" for c in new_schema["dataframes"][key]["columns"])
-            changes.append(f"new dataframe '{key}' added with columns: {cols}")
-        elif key not in new_schema["dataframes"]:
-            changes.append(f"dataframe '{key}' was removed")
-        else:
-            prev_cols = set(prev_schema["dataframes"][key]["columns"])
-            new_cols = set(new_schema["dataframes"][key]["columns"])
-            added_cols = sorted(list(new_cols - prev_cols))
-            removed_cols = sorted(list(prev_cols - new_cols))
-            if added_cols:
-                cols_str = ", ".join(f"'{c}'" for c in added_cols)
-                changes.append(f"'{key}' now has additional column(s): {cols_str}")
-            if removed_cols:
-                cols_str = ", ".join(f"'{c}'" for c in removed_cols)
-                changes.append(f"'{key}' had column(s) removed: {cols_str}")
-    return "\n# ".join(changes) if changes else "no change"
-
-
-def get_inp_op_metadata_comments(
-    parent_op: list[dict], ip: list[dict], op: list[dict]
-) -> tuple[str, str]:
-    def is_df_schema(d: dict) -> bool:
-        if not isinstance(d, dict) or not d:
-            return False
-        return "dataframes" in d and isinstance(d["dataframes"], dict)
-
-    def get_compact_summary(meta: dict) -> str:
-        if not meta:
-            return ""
-        keys_str = ", ".join(meta.keys())
-        return f"metadata with keys: {keys_str}"
-
-    input_descriptions = []
-    # Filter out empty dicts from the lists completely
-    parent_ops_clean = [p for p in parent_op if p]
-    ips_clean = [i for i in ip if i]
-    ops_clean = [o for o in op if o]
-    for incoming in ips_clean:
-        if incoming in parent_ops_clean:
-            continue
-        if is_df_schema(incoming):
-            input_descriptions.append(f"input: {describe_schema(incoming)}")
-        elif incoming:
-            input_descriptions.append(f"input: {get_compact_summary(incoming)}")
-    input_comment = "\n# ".join(input_descriptions) if input_descriptions else ""
-
-    output_descriptions = []
-    if ops_clean:
-        primary_op = ops_clean[0]
-        if is_df_schema(primary_op):
-            # Find the most relevant baseline to compare against (input or parent output)
-            baseline = next((i for i in ips_clean if is_df_schema(i)), None)
-            if not baseline:
-                baseline = next((p for p in parent_ops_clean if is_df_schema(p)), None)
-            if baseline:
-                output_descriptions.append(
-                    "output: " + compare_dataframe_schemas(baseline, primary_op)
-                )
-            else:
-                output_descriptions.append(
-                    f"output: new bundle: {describe_schema(primary_op)}"
-                )
-        else:
-            output_descriptions.append(f"output: {get_compact_summary(primary_op)}")
-    output_comment = "\n# ".join(output_descriptions) if output_descriptions else ""
-    return input_comment, output_comment
-
-
 def inputs_to_python(
     incoming_edges: list[workspace.WorkspaceEdge], saved_values: dict[str, str]
 ) -> list[str]:
@@ -314,39 +231,56 @@ def inputs_to_python(
     return inputs
 
 
-def workspace_to_python(ws: workspace.Workspace) -> str:
-    code = [
-        '"""The Python representation of the workspace."""',
-        "# All imports are handled automatically. Do not add imports.",
-        '# Only comments starting with "#!" are visible to the user and are interpreted as markdown. All other comments are for internal use only.',
-        "# Comments not separated by a non-comment line are grouped together into a single comment. For example, the following two lines:",
-        "#   #! This is a comment",
-        "#   #! This is part of the same comment",
-        "# will be grouped together into a single comment.",
-        "# Therefore, if you want to create a new comment, make sure to add a non-comment line in between.",
-        "# New user comments are placed above the next available function below it. (this rule doesn't apply to comments that are already in the workspace, they will stay where they are)",
-        "# For example, if a comment is on line 1 and on line 3 and the next function is on line 4, the both comments will be placed above the box represented by the function on line 4.",
-        "# Always place comments above the relevant line of code, so they appear above the box they are associated with.",
-        "",
-    ]
-    node_by_id = {node.id: node for node in ws.nodes}
+def _extract_info_from_edges(node_by_id, ws):
     incoming_edges: dict[str, list[workspace.WorkspaceEdge]] = {
         node.id: [] for node in ws.nodes
     }
-    outgoing_count: dict[str, int] = {node.id: 0 for node in ws.nodes}
     dependencies: dict[str, set[str]] = {node.id: set() for node in ws.nodes}
     for edge in ws.edges:
         # Ignore broken edges that point to missing nodes.
         if edge.source not in node_by_id or edge.target not in node_by_id:
             continue
         incoming_edges[edge.target].append(edge)
-        outgoing_count[edge.source] += 1
         dependencies[edge.target].add(edge.source)
     groups = {}
     for node in ws.nodes:
         if hasattr(node, "parentId") and node.parentId:
             dependencies[node.parentId].add(node.id)
             groups.setdefault(node.parentId, []).append(node.id)
+    return incoming_edges, dependencies, groups
+
+
+def _get_fnc_call(node, incoming_edges, saved_values, groups):
+    inputs = inputs_to_python(incoming_edges[node.id], saved_values)
+    params = sorted(f"{name}={repr(value)}" for name, value in node.data.params.items())
+    if node.type == "node_group":
+        params.append(
+            f"group=[{', '.join(saved_values[n] for n in groups.get(node.id, []))}]"
+        )
+    meta = node.data.meta
+    short_id = "".join(c if c.isalnum() else "_" for c in node.data.title.lower())
+    if meta and meta.python_function_name:
+        function_name = meta.python_function_name
+    else:
+        function_name = short_id
+    call = f"{function_name}({', '.join(inputs + params)})"
+    return function_name, short_id, call
+
+
+def _add_comment(node, code, saved_values):
+    comment_lines = node.data.params.get("text", "").split("\n")
+    saved_values[node.id] = (
+        f'"{node.data.params.get("text", "").replace("\n", "\\n")}"'  # node.data.params.get("text", "")
+    )
+    for line in comment_lines:
+        code.append(f"#! {line}")
+    code.append("")
+
+
+def workspace_to_python(ws: workspace.Workspace) -> str:
+    code = [WORKSPACE_PROMPT]
+    node_by_id = {node.id: node for node in ws.nodes}
+    incoming_edges, dependencies, groups = _extract_info_from_edges(node_by_id, ws)
     sorter = graphlib.TopologicalSorter(dependencies)
     sorted_node_ids = list(sorter.static_order())
     saved_values: dict[str, str] = {}
@@ -354,36 +288,17 @@ def workspace_to_python(ws: workspace.Workspace) -> str:
     for node_id in sorted_node_ids:
         node = node_by_id[node_id]
         if node.type == "comment":
-            comment_lines = node.data.params.get("text", "").split("\n")
-            saved_values[node_id] = (
-                f'"{node.data.params.get("text", "").replace("\n", "\\n")}"'  # node.data.params.get("text", "")
-            )
-            for line in comment_lines:
-                code.append(f"#! {line}")
-            code.append("")
+            _add_comment(node, code, saved_values)
             continue
-        inputs = inputs_to_python(incoming_edges[node_id], saved_values)
-        params = sorted(
-            f"{name}={repr(value)}" for name, value in node.data.params.items()
+        function_name, short_id, call = _get_fnc_call(
+            node, incoming_edges, saved_values, groups
         )
-        if node.type == "node_group":
-            params.append(
-                f"group=[{', '.join(saved_values[n] for n in groups.get(node_id, []))}]"
-            )
-        meta = node.data.meta
-        short_id = "".join(c if c.isalnum() else "_" for c in node.data.title.lower())
-        if meta and meta.python_function_name:
-            function_name = meta.python_function_name
-        else:
-            function_name = short_id
-        call = f"{function_name}({', '.join(inputs + params)})"
         parent_op_metadata = [
             node_by_id[e.source].data.output_metadata
             for e in incoming_edges[node_id]
             if node_by_id[e.source].data.output_metadata
         ]
-
-        input_comment, output_comment = get_inp_op_metadata_comments(
+        input_comment, output_comment = workspace_comments.get_inp_op_metadata_comments(
             parent_op=reduce(lambda x, y: x + y, parent_op_metadata, []),
             ip=node.data.input_metadata if node.data.input_metadata else [],
             op=node.data.output_metadata if node.data.output_metadata else [],
